+++ chat-widget/src/common/storage/default/defaultCacheManager.ts
@@ -1,5 +1,5 @@
 import { BroadcastService } from "@microsoft/omnichannel-chat-components";
-import { getWidgetCacheId } from "../../utils";
+import { StorageType } from "../../Constants";
 import { defaultClientDataStoreProvider } from "./defaultClientDataStoreProvider";
 
 export class defaultCacheManager {
@@ -7,15 +7,14 @@ export class defaultCacheManager {
     public static InternalCache: any = {};
 }
 
-export const registerBroadcastServiceForLocalStorage = (orgid: string, widgetId: string, widgetInstanceId: string, ttlInMins: number) => {
-    const widgetCacheId = getWidgetCacheId(orgid, widgetId, widgetInstanceId);
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+export const registerBroadcastServiceForStorage = (widgetCacheId: string, ttlInMins: number, storageType: StorageType) => {
     BroadcastService.getMessageByEventName(widgetCacheId)
         .subscribe((msg) => {
             try {
-                defaultClientDataStoreProvider(ttlInMins).setData(
+                defaultClientDataStoreProvider(ttlInMins, storageType).setData(
                     widgetCacheId,
-                    JSON.stringify(msg.payload),
-                    "localStorage");
+                    JSON.stringify(msg.payload));
             } catch (error) {
                 console.error("Error in setting data to localstorage", error);
             }