+++ chat-widget/src/components/prechatsurveypanestateful/PreChatSurveyPaneStateful.tsx
@@ -1,9 +1,8 @@
 import { HtmlAttributeNames, Regex } from "../../common/Constants";
 import { LogLevel, TelemetryEvent } from "../../common/telemetry/TelemetryConstants";
 import React, { Dispatch, useEffect } from "react";
+import { extractPreChatSurveyResponseValues, findAllFocusableElement, getStateFromCache, getWidgetCacheId, isUndefinedOrEmpty, parseAdaptiveCardPayload } from "../../common/utils";
 import MarkdownIt from "markdown-it";
-import { extractPreChatSurveyResponseValues, findAllFocusableElement, getStateFromCache, isUndefinedOrEmpty, parseAdaptiveCardPayload } from "../../common/utils";
-
 import { ConversationState } from "../../contexts/common/ConversationState";
 import { ILiveChatWidgetAction } from "../../contexts/common/ILiveChatWidgetAction";
 import { ILiveChatWidgetContext } from "../../contexts/common/ILiveChatWidgetContext";
@@ -66,15 +65,16 @@ export const PreChatSurveyPaneStateful = (props: IPreChatSurveyPaneStatefulParam
             dispatch({ type: LiveChatWidgetActionType.SET_CONVERSATION_STATE, payload: ConversationState.Loading });
 
             try {
-                const persistedState = getStateFromCache(state.domainStates?.telemetryInternalData?.orgId ?? "",
+                const widgetInstanceId = getWidgetCacheId(state.domainStates?.telemetryInternalData?.orgId ?? "",
                     state.domainStates.telemetryInternalData?.widgetId ?? "", state.domainStates.widgetInstanceId ?? "");
+                const persistedState = getStateFromCache(widgetInstanceId);
                 let optionalParams: StartChatOptionalParams = {};
 
                 //Connect to Active chats and chat is not popout
                 if (persistedState &&
                     !isUndefinedOrEmpty(persistedState?.domainStates?.liveChatContext) &&
                     persistedState?.appStates?.conversationState === ConversationState.Active &&
-                    state.appStates.hideStartChatButton === false) {
+                    state?.appStates?.hideStartChatButton === false) {
                     optionalParams = { liveChatContext: persistedState?.domainStates?.liveChatContext };
 
                     await initStartChat(optionalParams, persistedState);