+++ chat-widget/src/contexts/common/LiveChatWidgetContextInitialState.ts
@@ -2,21 +2,20 @@ import { ConversationState } from "./ConversationState";
 import { ILiveChatWidgetContext } from "./ILiveChatWidgetContext";
 import { ILiveChatWidgetProps } from "../../components/livechatwidget/interfaces/ILiveChatWidgetProps";
 import { defaultMiddlewareLocalizedTexts } from "../../components/webchatcontainerstateful/common/defaultProps/defaultMiddlewareLocalizedTexts";
-import { getWidgetCacheId, isNullOrUndefined } from "../../common/utils";
+import { getWidgetCacheIdfromProps, isNullOrUndefined } from "../../common/utils";
 import { defaultClientDataStoreProvider } from "../../common/storage/default/defaultClientDataStoreProvider";
-import { Constants } from "../../common/Constants";
+import { ConfirmationState, Constants, ConversationEndEntity, StorageType } from "../../common/Constants";
 
 export const getLiveChatWidgetContextInitialState = (props: ILiveChatWidgetProps) => {
 
-    const widgetCacheId = getWidgetCacheId(props?.chatSDK?.omnichannelConfig?.orgId,
-        props?.chatSDK?.omnichannelConfig?.widgetId,
-        props?.controlProps?.widgetInstanceId ?? "");
-
+    const widgetCacheId = getWidgetCacheIdfromProps(props);
     const cacheTtlInMins = props?.controlProps?.cacheTtlInMins ?? Constants.CacheTtlInMinutes;
-    const initialState = defaultClientDataStoreProvider(cacheTtlInMins).getData(widgetCacheId, "localStorage");
+    const storageType = props?.useSessionStorage === true ? StorageType.sessionStorage : StorageType.localStorage;
+    const initialState = defaultClientDataStoreProvider(cacheTtlInMins, storageType).getData(widgetCacheId);
 
     if (!isNullOrUndefined(initialState)) {
-        return JSON.parse(initialState);
+        const initialStateFromCache: ILiveChatWidgetContext = JSON.parse(initialState);
+        return initialStateFromCache;
     }
 
     const LiveChatWidgetContextInitialState: ILiveChatWidgetContext = {
@@ -34,13 +33,16 @@ export const getLiveChatWidgetContextInitialState = (props: ILiveChatWidgetProps
             customContext: undefined,
             widgetSize: undefined,
             widgetInstanceId: "",
-            initialChatSdkRequestId: ""
+            initialChatSdkRequestId: "",
+            transcriptRequestId: "",
+            confirmationPaneConfirmedOptionClicked: false,
+            confirmationState: ConfirmationState.NotSet
         },
         appStates: {
             conversationState: ConversationState.Closed,
-            isMinimized: false,
+            isMinimized: undefined,
             previousElementIdOnFocusBeforeModalOpen: null,
-            isStartChatFailing: false,
+            startChatFailed: false,
             outsideOperatingHours: false,
             preChatResponseEmail: "",
             isAudioMuted: null,
@@ -54,11 +56,9 @@ export const getLiveChatWidgetContextInitialState = (props: ILiveChatWidgetProps
             },
             e2vvEnabled: false,
             unreadMessageCount: 0,
-            conversationEndedByAgentEventReceived: false,
-            conversationEndedBy: undefined,
-            postChatWorkflowInProgress: false,
-            shouldUseBotSurvey: false,
-            chatDisconnectEventReceived: false
+            conversationEndedBy: ConversationEndEntity.NotSet,
+            chatDisconnectEventReceived: false,
+            selectedSurveyMode: null
         },
         uiStates: {
             showConfirmationPane: false,