+++ chat-widget/src/index.ts
@@ -2,7 +2,9 @@ import { BroadcastService, decodeComponentString, encodeComponentString } from "
 
 import useChatContextStore from "./hooks/useChatContextStore";
 import useChatSDKStore from "./hooks/useChatSDKStore";
-
+import { getWidgetCacheId, getWidgetEndChatEventName } from "./common/utils";
+import { ConversationState } from "./contexts/common/ConversationState";
 export { default as LiveChatWidget } from "./components/livechatwidget/LiveChatWidget";
 
 export { encodeComponentString, decodeComponentString, BroadcastService, useChatSDKStore, useChatContextStore };
+export { getWidgetCacheId, getWidgetEndChatEventName, ConversationState };